Implement artifactless conduct
This conduct is broken whenever you touch an artifact, which is not
possible to do by accident. The only ways to break this conduct are to
wish for or pick up an artifact, to dip for Excalibur, or to get
Excalibur during lawful crowning (neutral and chaotic crowning drop the
weapons at your feet).

Note that the unique items (Bell, Book, Candelabrum, Amulet of Yendor)
are not artifacts and the player may touch and use them freely without
breaking this conduct.