Generate many more monsters when level is created, and fewer over time
Normal room-and-corridor levels now have a 1/2 chance of placing a
monster in each room, and when a monster is placed in a room, there is a
recursive 1/3 chance to place another, and so on. Maze levels now place
rn1(10, 20) random monsters instead of rn1(5, 7).

This is to compensate for a large decrease in the monster spawn rate.
Monsters spawn only on 1/150 of turns (down from 1/70) when above the
Castle, 1/100 of turns (down from 1/50) when below it, and 1/40 of turns
(down from 1/25) when u.uevent.demigod is on.