Replace poison instadeath with severe HP loss
Unavoidable poison instadeath is removed. Instead of poison being
"deadly", it will now be "potent", and will cut your current HP to half
and your maximum HP to two-thirds, as well as draining an attribute.

Poison instadeath on the monster side is also removed. Monsters' HP is
now reduced to 1 instead of killing them outright.