Add jonadab's "Gnomish Sewer" Mines' End level variant
The Gnomish Sewer is a level filled with waterways and adjacent paths.
These paths are irregular and navigation through the level can be very
difficult without the proper equipment. A few parts of the level are
completely inaccessible without digging or squeezing through diagonals.
There are a multitude of monsters, but most will probably not be
particularly threatening except for the three electric eels, two
crocodiles, and possible purple worm and chameleon. Between the terrain,
the monsters, and the fact that the luckstone might be anywhere on the
level, this will likely be the hardest Mines' End variant.

To compensate for this a bit, there are seven random rings scattered
around the level (having come down the pipes from being dropped down
sinks) and a touchstone as well as the usual assortment of gems. There
are no specific hard gems here; instead there is one of each type of
soft gem, and two random gems.

Differences from the original proposal:
 - A couple spots are floor areas instead of corridor.
 - There are 6 rust traps and 3 random traps instead of 3 and 6.
 - Add two gnomes and 1-2 each of brown puddings and gray oozes.
 - Add a blank scroll, randomly placed.