Implement FIQHack's Pw regen formula
This changes the power regeneration formula to a very slightly modified
version of FIQHack's. The main change behind the scenes is that energy
(but not maximum energy) is now a float. In most cases (modifying and
comparing energy to other quantities) nothing has needed to be changed,
but in a couple cases like botl display it is treated as the floor of
its current value.

The new formula for power regeneration, from FIQhack, is to stack all
the following together:
- 1 Pw per turn if you have energy regeneration
- 3*XL/100 Pw per turn
- 3*Wis/100 Pw per turn
- 33/100 Pw per turn if Wizard

Because energy is now a float, there's no need for a complicated formula
to determine whether to regenerate a point on this turn; fractional
quantities are added directly.