Add a shop and reasonable gear to Orcish Town
This is an attempt to address the massive number of complaints with
Orcish Town, namely that it has few things of use, by resurrecting its
general store with a living shopkeeper.

The gameplay reason is to give the player a means to price ID and
several normal random objects to buy. The flavor is that the shopkeeper
barricaded himself in his shop with boulders before the orcs could reach
him. The player will need some way to break the boulders or dig the
surrounding terrain to access the shop.

Additionally, some of the loot one would expect from the Watch and the
priest is added in as well; armor, weapons, and whistles for the Watch,
a possible key from a shopkeeper, and a robe, shield, and spellbooks for
the priest. None of these are guaranteed, but at least some should
generate in any game.