MS_BRIBE monsters now demand 50d500 gold
This change is aimed at making gold actually matter more in the endgame,
with demon lords demanding a high but widely variable amount of gold, so
the player won't be completely safe unless they have and are willing to
forfeit 25,000 zorkmids.

There is a 1/50 chance that, if the player can't meet the demand but
surrenders all their gold, the demon will accept the bribe anyway.

The previous shortchanging formula (when the demon scowls menacingly,
then vanishes) works the same way, but the margin has been expanded from
40 to 2000.