Add graveyard to FLAGS for Priest quest and Castle
All special levels of the Priest quest and the Castle are considered
graveyard levels for the purposes of undead corpse generation, etc, but
were hardcoded to set that flag in mkmaze.c. I can see no point in
keeping them as special cases in the code if they can be specified in
the des files, so this commit adds them to the FLAGS of these levels and
removes the special cases in mkmaze.c.

The Priest quest filler levels do not have FLAGS defined, but they have
guaranteed graveyards, which will make sure that the graveyard flag
continues to be set.