Wielding a mirror grants reflection half of the time
If the hero has no other source of reflection but needs to defend
against something that can be reflected, a mirror may work in a pinch.
Half of all reflectable incoming attacks will be reflected if the player
is wielding a mirror.

As part of this change, the ureflects() function is removed and replaced
by a ureflectsrc() function; the old function's job of printing various
messages is removed and the various callers print the messages instead.