Move lizards from : to l, and merge leprechauns into i
This is an effort to move a monster class off a punctuation mark. In
vanilla, l is a one-monster class so it's a good candidate to move
lizards to. Leprechauns are similar enough to imps and homunculi that
they can be placed in i. This changes the class description of i to "an
impish creature".

This commit was recreated from the old repository because git wouldn't
apply the patch properly.