Merge ghost monster class ' ' into W
Ghost monster class being a blank space is very annoying, and ghosts are
wraith-like monsters, so the classes are being merged here. To determine
whether something is a ghost, instead of comparing its symbol to
S_GHOST, use noncorporeal().

Color changes include:
- Ghosts are white and shades are black.
- Wraiths are now gray.
- Barrow wights are now bright blue for spellcasting.

This commit was re-created from the old repository because git is being
stupid.