Pets can get satiated and won't eat food when satiated
This is from either Fourk or FIQHack. If the pet's internal 'hungrytime'
(the turn on which it will get hungry again) is more than 1500 turns
in the future, it will not eat anything, even treats. This means if you
keep your pet well-fed, it will not try to nab that floating eye corpse
away from you...

I changed the existing DOG_SATIATION from 3000 to 1500 after seeing the
EvilHack reasoning for the same: with pet intrinsics, having too much of
a satiation buffer might make it hard to set up situations to give your
pet intrinsic-granting corpses.