Flavor non-spiked pits in graveyards as open graves
When the player falls into a pit in a graveyard, it will print messages
relating to it being an "open grave" instead of a pit. Additionally,
there are two messages added: "It's a bit early for that, isn't it?" if
the player is living, and "It seems quite cozy down here." if the player
is undead.