Rocks can be broken to produce flint stones, and flint weighs 4
The player can apply or rub rocks on other rocks or gray stones to bang
them together, breaking the rock and possibly creating some flint
stones. The number of flint stones produced is usually 0, but it helps
boost the chances of getting one or more if you are rubbing the rocks on a
touchstone or you are a Caveman.