Complete door overhaul
This is a large commit that overhauls the door bitmask completely,
separating it into discrete 'states' and 'flags'. States occupy the low
2 bits, and represent what the actual state of the door is, regardless
of any other modifiers. It can be no door, broken, open, or closed.
Flags occupy the 3rd, 4th, and 5th bits, and represent independent
characteristics that may or may not be applied to a given door. The 3rd
bit is locked, and the 4th is trapped; both of these are only meaningful
on closed doors. The 5th bit is unused as of this commit, but will be
used for iron doors in the future.

Also, I have added a bunch of macros for accessing and modifying doors,
and nearly all places in the code have been changed to use them. The
D_NODOOR, D_BROKEN, D_OPEN and D_CLOSED should essentially only be used
in conjunction with doorstate() or set_doorstate(). D_LOCKED, D_TRAPPED
and D_WARNED should not have to be used anywhere in src/.

There are some changes to the level compiler as well:
 * If you want to specify a container to be trapped, use 'ctrapped'.
 * Door state can be specified several ways:
   * Simple door state: DOOR:closed,(0,0)
   * Just locked or trapped flags: DOOR:locked,(0,0) - this forces the
     state to be closed, DOOR:locked&trapped,(0,0) also works
   * State plus flags: DOOR:closed&trapped,(0,0)