Replace spell failure with increased Pw cost
This removes the mechanic in which you "fail to cast a spell correctly"
through some random failure percentage chance. (It will still happen
when confused.)

Instead, the percentage chance of failure is used to increase the Pw
cost of the spell. The increase is inversely proportional to the chance
of success: the Pw cost is the same as before if the success rate was
100%, doubled if the success rate was 50%, quadrupled if the success
rate was 25%, and so on. If the success rate was 0%, the Pw cost is
infinite and you will not have enough energy to cast it no matter what.

The spell menu no longer shows Fail% numbers, it shows the direct Pw
cost instead, or "Inf" if infinite.

This also adds a function energy_cost() that computes the Pw cost of a
given spell.