Amulet of life saving does not work if the player is nonliving
This was already the case for a monster that happened to be wearing
such an amulet. On the player, it should currently do nothing
different, but this will change if the player is ever able to get
into a state where they are nonliving, won't rehumanize, and are wearing
an amulet of lifesaving instead of unchanging (planned).