The game might continue when you would rise from the grave
Certain monsters kill the player and reanimate their corpse, or turn
them into a monster which, evidently, loses its mind and becomes an
ordinary (albeit named) monster in the bones file.

In a fraction of cases, such deaths now do not actually kill the player;
instead, the player's body corrupts and transforms into a monster of the
appropriate type, but the player remains in control. The drawback to
this happening is that the player now gains permanent intrinsic
unchanging; the rest of the game must be played out in that form.

If the player is unchanging (whether from unluckily wearing an amulet or
from having died and transformed in this manner already), this effect
does not happen.

Also, a tweak had to be made to the lifesaving logic: if you are in
creature form with unchanging and die, lifesaving will only restore your
full monster HP; it will _not_ rehumanize you.