object probabilities can add up to different values than 1000
Allows object probabilities to be tuned with arbitrary accuracy, rather
than in 0.1% increments. In particular, when adding a new item, one does
not have to steal probability from other items to make it add up to the
same total. Probability can now be thought of as (prob / sum of probs),
rather than (prob / 1000). This is computed on-the-fly when generating
an object.

This also removes makedefs and other checks whose purpose was to make
sure all probabilities in a class added up to 1000. It does not,
however, touch the code specific for setting gem probabilities based on
dungeon level (whether probabilities still add up to 1000 or not, this
new system will still choose an object proportional to its probability.)