Add thiefstone and make it wishable
Trivial addition to objects.c for thief stone, no behavior implemented
yet.

Wish parser changes are more significant: adds "thief stone" as a valid
alternate spelling, but it also has to special-case "thiefstone" in the
code that checks for monster names (because "Thief" is a Rogue title).
Ended up moving the code that checks for alternate spellings (and jumps
past the monster-name-checking) above the monster-name-check code.
Hopefully this doesn't break anything; tested with things like dragon
scales and cans of specific monster meat, and it all appears to work.