Implement thief stones keying to a specific location
Thiefstones still do nothing, but they will now correctly choose a keyed
location when they are first created. This will always select a square
in a vault if there is one, and will otherwise use a heuristic to choose
from among all squares on the level, preferring untrapped floor or
corridor spaces, closets, and spaces with containers.

The keyed level is stored as a ledger number using the spe field, and
the keyed coordinates are stored in the lowest 8 and next-lowest 8 bits
of the corpsenm field. Perhaps an improvement would be to associate a
coordinate and d_level in the oextra struct, but currently no other
objects would use them.

Thiefstones can also be cancelled; this is whenever the ledger number is
0 (normally illegal). A normal thiefstone will display (keyed to
Branch:depth) when identified; a cancelled one will just display
(inactive). Wizard mode will additionally show the keyed coordinates.

This also defines a niche using level flags (so that thiefstones can pick
up on it). is_niche will be set to 1 only when a closet is specifically
generated.

Other changes include: thief stones now count as gray stones for the
purposes of is_graystone, the ROCK definition in objects.c is now above
the actual rocks, and it has been expanded to include the mrg and uskn
flags. All ROCK objects previously had mrg=1 and uskn = 0, but thief
stones require mrg=0 and uskn=1.