Changes to migration code in preparation for thiefstones
Adds two fields to struct obj, mgrx and mgry, to hold the x and y
coordinates, if known, of an object in migration. Also adds an alias for
owornmask named "migrateflags" to remove ambiguity about when it's being
used for migration and when it's being used as a worn mask.

Also adds a new migration code, MIGR_THIEFSTONE.