Fully implement thiefstone item teleportation
Thiefstones now teleport their chosen item to the correct square on the
correct level. For levelporting items, the objects are migrated. Two new
fields had to be added to struct obj, mgrx and mgry, to save the exact
position of a migrating object. This also implements MIGR_EXACT_XY for
migrating objects as a side effect, but thiefstone teleports use a new
MIGR_THIEFSTONE that not only delivers the items to the exact spot but
also will put them into a container found on that spot, if any.

Also adds a thiefstone_accept function that defines which items will be
teleported by thiefstones (never worn items, only oc_magic items, except
gold and gems including glass if the thiefstone is blessed.)

Also implements the second case of involuntary thiefstone teleport, when
the hero tries to pick up a cursed thiefstone and it pulls some
acceptable item out of the pack and teleports it.

Lastly, adds a couple other things to the init_thiefstone function: it
will only create a chest if there isn't already one there, and it will
always generate a thiefstone as cursed.