Blessed thief stones now teleport gold golems to the keyed location
Only blessed ones, since they are the only ones that accept gold as a
valid target. They don't follow the usual hit logic - the player needs
to throw the stone at the golem to get it to hit (tried implementing it
in hmon_hitmon so melee would also work, but couldn't fix the panics
that resulted in without a major refactor.) So the only caveat is that
wielding a thiefstone and hitting a gold golem with it will not work.

The player can take advantage of this to return to the keyed location by
polyselfing into a gold golem and throwing the stone upwards - it will
then "steal you away!", sending you to the keyed location.