Implement functions for door trap variety
This adds two functions: getdoortrap and doortrapped, which determine
door trap behavior chosen from a pool of possible traps instead of a
guaranteed stunning explosion. Door traps are not stored in any level
structs, but are rather selected deterministically based on x, y, level,
and birthday. getdoortrap returns the type of trap; doortrapped defines
the behavior for when a player triggers a trapped door.

New types of door traps range from screeching hinges to dumping a bucket
of water on the player's head to the door falling off the hinges. Each
trap has a minimum level on which it can appear.

Other changes: adds TRAPPED_DOOR to the list of psuedo-objectclasses
that can be passed to explode(), the death message for being hit in the
head is now "crushed by a falling boulder" instead of "scroll of earth".