Extend doortrapped to handle monster cases
Monsters are affected as symmetrically as possible by door traps. The
various calls to b_trapped and mb_trapped are only partially replaced
here, so there are currently no cases where a monster will trigger
doortrapped code except by blasting down a door with a wand.

doortrapped now takes six arguments: an x and y coordinate, pointer to
the triggering monst, the body part, the action, and a number indicating
whether this trap is triggered before, after, or regardless of when the
interaction with the door happens. Also provides macros predoortrapped()
and postdoortrapped() so this sixth argument doesn't need to be used
directly in most cases.

The places in the code that now contain doortrapped() calls are picklock
and doorlock. doorlock had to be extended to take a monster responsible
argument, since doortrapped wouldn't otherwise know.

doortrapped's ability to return whether the trap destroyed the door or
not is helping clean up a lot of duplicate code around trap calls, such
as manually unblocking the point and charging the player for shop door
damage when the door is destroyed by a trap.