Implement pouring potions down sinks
Allows for safer identification of potions by dumping them down the
drain, leaving the hero to suffer only vapor effects (none of which have
been changed yet). This is done by selecting a potion to #dip while
standing on top of a sink.

The exception to vapors is the potion of polymorph, which will polymorph
the sink instead, the same as dropping a ring of polymorph into it will.