Rebalance spell book failure effects
Addresses balance problems I saw with failure effects that discourage
spellbook reading. Most importantly, the side effect that accompanies
all failures has been changed from paralysis of (duration) turns to
confusion of (duration) turns. Also, an effect that can happen at any
level is for the book to become unblessed or cursed and glow black.
Reading a cursed book is now the only way for the book to disintegrate.

Other effects are:
  Level 1+: If cursed or with 75% chance, teleport; else unbless/curse
  Level 2+: aggravate monsters (unchanged)
  Level 3+: blind for rn1(30, 10) turns (duration reduced)
  Level 4+: book bites you for rn1(5, 3) damage
  Level 5+: book paralyzes you for rn1(16, 16) turns
  Level 6+: contact poison (unchanged)
  Level 7: exploding rune (unchanged)