Some more spell changes: no conf paralysis, BUC is +/- 10 Int
Removes paralysis from becoming confused while reading a book, and also
when trying to read a book while confused but not triggering cursed
book. It is no longer possible to die with the multi_reason of "reading
a book".

Also makes the change that blessed spellbooks aren't guaranteed to
succeed, and cursed ones aren't guaranteed to fail. Instead, the
beatitude acts as a +10 or -10 to the player's effective Int.