Allow the player to wish for a magic lamp
I consider this to be acceptable since the player is forfeiting a 100%
probability of a wish right now for 80% probability of a wish later. It
can theoretically be reused to generate multiple oil lamps, but this
will eventually end when a djinni refuses to give a wish, and the light
from a magic lamp is infinite anyway.

This also allows players to spend a wish from something that isn't a
magic lamp on a magic lamp so that they have an infinite light source.