fix impossible() in object lookup
If the player tried to look up a bare object class like "ring",
name_to_otyp would parse past it resulting in an empty string, and then
fruit_from_name would try to singularize it. Now, if the input is
reduced to an empty string, it will immediately return no otyp.