Make useless +0 rings impossible to generate
Magic rings are pretty much useless to most characters if they are +0;
they are generally not worth the charging resources to enchant them up.
This removes the 10% chance that they'll generate at +0, forcing them to
generate at either negative or positive values with what is currently
equal chances. (The 80% chance of a ring that generated uncursed and
negative becoming cursed is kept.)