Add a see_monsters() call to fix monsters being overwritten with gold
With this, passive gold detection should pretty much be complete.
There's a possible minor bug in that the detection will map object piles
which should perhaps not be shown, but it's not clear whether or not
that should be fixed.

The only other concern is performance; a character with passive gold
detection will force a newsym() iteration over the entire map per every
action, as well as an iteration over the level objects list and then two
see_monsters calls. On my system there's no noticeable difference, but I
can imagine it causing issues on older systems.