Standardize all roles' "cutoff levels" for maxHP/maxPw increases
They varied quite a bit, from 10 for several classes all the way up to
20 for healers. This removes the xlev member from struct Role entirely
rather than just editing values to 12.

I may in the future want to strike out the entire system of having
different formulas for maxHP and maxPw increases, but finding a formula
that will work with the Con/Wis modifiers may be tricky. In general a
formula that decreases the amount of randomness and giving roles a
larger proportion of fixed points would be preferred.