Remove the spell of identify outright
FIQ pointed out that having it always identify one item would not
actually make scrolls of identify preferred, it would just stretch out
the ID game even more and make everything more tedious. I think that as
long as there is a renewable source of type-identification, the ID game
will remain effectively broken for any character that becomes capable
of casting identify, and it's hard to justify the spell of identify if
it can't do type-identification.