Remove b_trapped from the game
With all instances of door traps converted to doortrapped() calls, the
only remaining instance was tin traps. The bulk of the b_trapped code
has moved into there and the function is removed since there's nothing
that calls it anymore.

Note that mb_trapped still exists as of this commit, but removing that
is next.