Replace hash function that determines door trap
The previous off-the-cuff "hash" function was severely statistically
skewed, and wouldn't produce an even distribution of types of door traps
per level. It also had noticeable artifacts such as a single row or
column only generating certain types of traps. This replaces it with one
that generates a unique assignment for a coordinate based on x, y,
ubirthday, and ledger_no (it previously used level difficulty but this
would mean that the same coordinates on two different levels with the
same difficulty would have the same trap).  Then this number is hashed
into oblivion and the final trap selected based on it.