Implement monsters interacting with doors and door traps
Monsters now take a turn to interact with (any one of: unlock, open, or
smash down) doors, without moving. This was not previously a priority,
but the code previously assumed that a monster interacting with a door
would always move onto the door's space (unlock+open, open, smash down,
trigger a blow-up-door trap) which doesn't work anymore with the new
door traps (particularly self-locking doors).

Moves the former mon-interacting-with-door logic into a mon_open_door
function which determines whether a monster should be interacting with a
door at a given space to which it wants to move. The reason this needs
to be a function is that it also needs to get called by the priest and
shopkeeper move_special function and by the pet dog_move function, in
addition to m_move where the logic originally lived.

mb_trapped continues to exist, and the next step will be addressing
whether I've neglected any code paths before removing it entirely.