Trigger door traps when knocked down by a force bolt
Possible thing to be concerned about: zap_over_floor carries no
knowledge of who is responsible for the zap, so make sure no door traps
exist that would hold the player responsible for any monsters that died
as a result. As far as I know there aren't any such now.