Small gremlin fix; gremlins get angry when hurt with light
The blessed scroll of light hitting every square on the level produced
"It cries out in pain!" for out-of-sight gremlins anywhere on the level.

Also, gremlins now get angry at the player when they take damage from
light (since the player is the only source of damage dealt this way).