A few more recolors
The "orc" monster is recolored to blue to bring it in line with the
color of orc mummies and zombies. To avoid conflicting, Mordor orcs are
changed to red, which makes more sense given they serve the Red Eye
anyway.

The "giant" monster is recolored to cyan to bring it in line with the
color of giant mummies and zombies. This conflicts with hill giants, but
giants are unlikely to be encountered much and neither they nor hill
giants have any special abilities.

Acid blobs are recolored to bright green because it seems a better
"acid" color. (To be totally consistent, green molds and lichens should
then swap colors, but it would probably throw off too many players.)