Move bugbears into the o class
Because bugbears in D&D are a variant of orcs, they deserve to go on the
same glyph. Recolored to orange so as not to conflict with hobgoblins.

Caveman quest levels now generate random o-class monsters instead of
random humanoids, which is more flavorful anyway.

The random h in the mines now has a slightly higher chance of turning
out as a mind flayer, I suppose.