Wielding wands that correspond to spells
When you wield a wand that corresponds to a specific spell, you get the
same benefit as was added for wands of nothing and quarterstaves
(equivalent to a robe, while not stacking with it), but for that
specific spell only. The practical effect, given other xNetHack changes,
is that the Pw cost of a spell will be reduced to closer to its
baseline.

Casting the corresponding spell does not drain charges from the wand or
anything like that.

This may not be powerful enough, since a robe, quarterstaff or wand of
nothing makes it totally irrelevant, and it may be necessary to make it
even stronger than just being equivalent to them.