Full replacement of percent_success()
Replace the vanilla formula with one written by FIQ, with some slight
modifications. The formula is notable for being independent of role
stats, relying only on Int, XL, skill, armor, and which spell is being
cast. Additionally, it's rather easy to follow, since all of the inputs
directly affect the percentage points.

The formula is as follows:
1.  +5% per point of Intelligence.
2.  +5% per experience level.
3.  If wearing a robe, wielding a quarterstaff or wand of nothing, +50%.
4.  Otherwise, if wielding a wand that matches the spell, +70%.
5.  -25% per spell level.
6.  -50% for metallic body armor.
7.  -50% for wearing a bulky (oc_large) shield, or -15% for non-bulky.
8.  -20% for a metallic helm that isn't a helm of brilliance.
9.  -35% for metallic gloves.
10. -10% for metallic boots.
11. +20% for Basic skill in the spell school, +40% Skilled, +80% Expert
12. Clamp to the [0%, 100%] range.

This does not remove the now unused role statistics. If anyone wants to
remove these later, they are: spelbase, spelheal, spelstat, spelshld,
spelspec, spelsbon. spelarmr is NOT unused: it appears in uhitm.c for a
Monk-specific modifier, which could probably be hardcoded. Possibly, one
stat could be added back in for role differentiation purposes, if this
makes it too samey.