Make several changes to the Gnomish Sewer
A lot of places on the level were inaccessible without either water
walking or dropping lots of stuff to squeeze through a choke point. Now,
all the non-water squares on the map should be reachable from any other
non-water square without needing either of these.

Monsters are also changed: three electric eels plus the chance of 6+1d3
more random ; producing eels is too many. There is now one electric eel
who spawns on top of the wand of cold, and the 6 random ; have been
converted into piranhas and jellyfish. (Random ; are still possible in
the 1d3 loop, but only with a 25% chance each time.) One snake is
replaced with a water moccasin. The purple worm is much less likely to
occur, and the baby purple worm more likely.

There are now 2 random scrolls of blank paper around the level, and the
guaranteed random loadstone is removed.