Implement paranoid swim, encapsulating both water and lava
Adds "swim" to the list of paranoid options. The mechanics of this are
more or less ported directly from UnNetHack, where you will not be asked
to confirm entering liquid if you are levitating, flying, wearing known
[fireproof] water walking boots, moving with the 'm' prefix, etc.

As with all the other paranoid options, this is settable via config
file, as "swim".

Since having no confirmation for walking on lava means you must have
fire resistance, there's a minor but extremely marginal exploit in
identifying the ring of fire resistance. See the FIXME in this commit.