Implement paranoid trap
Port paranoid_trap from, again, UnNetHack to the 3.6.1 paranoid system.
Unlike paranoid_swim, this will always prompt before stepping on a trap
(unless stunned or confused).

Unimplemented as of yet, but planned, is disabling the confirmation when
you are immune to the trap (known magic resistant for polymorph,
levitating for hole, etc.)