Refactor immune_to_trap
Now returns either 0, 1, or 2; the last is a special value that
indicates that the player is in fact immune to the trap but may not
actually know that they are immune (due to unidentified items or hidden
intrinsics): in which case the interface should still warn them about
the trap.

Not a perfect implementation, because ideally if the player is wearing
an identified ring of fire resistance and carrying no burnables, then
they should be able to walk onto a fire trap, but good enough for now.