Remove random chance of extra-powerful poison entirely
The original attempt to ameliorate the poison instadeath by replacing it
with "potent" poison was unsatisfying and asymmetric. The player's HP
would be cut in half and their HP max reduced to 2/3, whereas monsters
would be reduced to 1 HP (in an effort to keep it worth it to use
poisoned weapons that no longer instakilled monsters.)

Since then I've come to think that a random chance of the poison
delivering an extra-powerful dose is rather arbitrary anyway, and have
decided to remove it completely. Poison may still hit the hero's
attributes as it did before, but in all other cases (monster poisoning
and hero non-attribute poisoning) it will do d6 extra damage if a thrown
weapon, and d10+6 damage from another source (sting, etc.)