Refactor: grounded() permonst macro
A lot of places in the code assume that (is_flyer() || is_floater()) is
a sufficient check for "monster is not touching the ground"; however,
this ignores clinging monsters like the wumpus. Some checks recognized
the wumpus, but hardcoded it specifically and ignored other clinging
monsters.

This creates the grounded() macro (meaning !is_flyer || !is_floater ||
!is_clinger), and replaces a lot of the old partial checks for
ground-bound monsters with it.

(Minor extension of this commit: it also creates an is_bee() macro that
returns true only for killer and queen bees.)