Change behavior of paranoid swim and interactions with liquid
Now, trying to step onto water or lava while you're not sure it would
affect you simply doesn't work at all and your move will be canceled.
You can deliberately attempt to move onto it with m. If paranoid swim is
turned on, you will then be asked to confirm if you want to enter the
water (with a yn prompt, not a paranoid confirmation). If off, it will
simply let you move onto the space (which I believe is NH4 behavior.)