Water walking boots self-identify when they prevent you entering water
Slight refactor of drown() so that the caller should not check water
walking first. Instead, drown() handles this check, and if the hero is
prevented from entering water by their water walking boots, the boots
are type-identified.

Also true for lava: if the hero walks on lava with fireproof water
walking boots and survives, the boots are identified along with their
fireproof state.