Refactor door generation code, traps can now generate on open doors
Remove the DOOR/SDOOR split with its different probabilities of locking
and trapping between regular and secret doors. Secret doors are now
special-cased only by forcing them to be closed, and by preventing them
from generating mimic traps.

Certain traps can now generate on open doors as well as closed.
(Currently there aren't any traps that would trigger on broken or absent
doors.)

Mechanical changes of this commit also include frequency adjustments:
locks appear on doors with 1/4 frequency instead of 1/6, for an overall
frequency of 1/15 (the previous frequency was 2/45). Traps appear on
doors with 1/23 frequency, but only if the door has already chosen a
state that is suitable for the type of trap at that location. Traps will
now convert into mimics with 1/7 frequency, versus 1/5.