Extend water_damage_chain; nerf effects of water bucket trap
Several players complained about the water bucket trap affecting every
single item in their inventory because it used water_damage_chain - in
one case, getting into a bag and blanking 70 scrolls.

The first thing this commit does is extend water_damage_chain to take 2
more arguments. The first argument represents the number of items in the
chain to damage. The second represents whether it should damage
containers. If the caller wants to specify a certain number of items to
be damaged, the function uses reservoir sampling to select that many
items randomly and with equal probability from the chain.

The water damage done by the bucket trap now specifically ignores
containers, and it will wet only (level difficulty/5) + 1 items.