Engravings display visibly on the floor
They will render as ", in various colors depending on what the type of
engraving is (brown for dust, gray for engraved, black for burnt, red
for blood, and bright green, bright blue, or bright magenta for
graffiti.

The current implementation is somewhat hacky - they copy AMULET_CLASS,
which is all right for ASCII users but will be very confusing in a
tileset. I intend to define a S_engraving glyph, but am not sure whether
this will savebreak, so for now it will remain a duplicate of amulets.