Spiders occasionally spin webs while moving around
In the m_move function, a check is made for if the monster is a
webmaker. If it is, and there is no existing trap on its space, it may
spin a web there. The probability of this is 2.5% per move for giant
spiders, and 1% for cave spiders or any other webmaker. The chance is
doubled if the spider is in a doorway (because spiders commonly spin
webs in doorways in real life).

If the player observes a spider spinning a web, the web is automatically
made visible.