Select gray stone locations in the Gnomish Sewer from a list
It's probably too cruel to have them be completely random, since it's
rather hard for a player to see whether they've explored all the twisty
passages. Make a list of random far-flung locations and use that.

Also tweak a couple of map tiles to connect an inaccessible passage and
connect a water passage to the starting room.