Refactor: allow get_rnd_text to select a specific line
Needed for tin labels, which must choose their label deterministically,
and one label requires getting a hallucinatory monster, which must also
be deterministic so get_rnd_text() as it exists doesn't cut it. Adds a
third argument to this and bogusmon(); if this third argument is
negative it will select the line randomly as normal; otherwise it will
choose the (X % (lines in file))th line of the file.