More library functions: single number hash, uppercase all words
The single number hash, hash1(), is a wrapper around coord_hash that
passes 0 as its x and y arguments, and is used when you have a single
number you want to hash.

The up_all_words() function turns the first letter in a string after
each space to uppercase. This is used for formatting tin label monster
names. It might also be useful later if we wanted to implement
Pratchett-Style Golem Speech.