Implement tin label reading
As with the other non-scroll non-book items you can read, tins now have
readable labels. This violates illiterate conduct and does not work
while blind. A few random tins come unlabeled, and any player-made tins
will also be unlabeled.

In an attempt to make tin labels worth reading, 1/5 of tins tell the
truth about their contents - not to the level of telling the player the
preparation method, but being honest about the creature inside. All
other tins have false labels, which can claim to be a monster that they
are not, claim to be a genocided or bogus monster, claim to consist of
the player's fruit, or show any of a number of obligatory tin-based
references and joke labels.