Polymorph traps disappear when they successfully polymorph something
This was inconsistent; the trap disappeared when the player set it off
but not when a monster did, which enabled players to pursue strategies
like repeatedly displacing a pet onto a polymorph trap while protected
themselves via magic resistance.

However, that's not the primary motivation for this change; the main
reason is so that a crowd of monsters can't all polymorph repeatedly
into too-strong forms. Also so that the player is in general less likely
to run into a polymorph trap in an area that monsters have obviously
passed through.